Fixed: Blood Demon Reset
Blood Demon's ultimate skill "Reset" was reported to not work, however it did actually work.
The suspected reason for the confusion was due to lack of feedback from the ultimate skill when it failed, so now whenever a player tries to use the ultimate it will not go on cooldown if the player has yet to kill an enemy.