Added: Item Mole
Added the item 'mole' to the shopmenu.

In the past a lot of server owners chose to disable this item from their shop, hence it's default state is set to disabled.

To enable the newly implemented mole item, simply go to cfg/source-python/wcs/items.ini, scroll down in the Enhancements section and find the mole item, remove all the double // in front of the item and it will be enabled on your server.

I hope you'll all enjoy the mole's return to the shopmenu! (^.^)