Fixed: Item Mole
Fixed the shop item 'mole', apparently I somehow managed to remove the es_xdoblock prior to uploading the es_mole.txt to github, quite a feat. This fixes the probem with mole not working.

Also made a slight change in the items.ini allowing players to buy the mole item while they are dead.