Fixed: Spawn Health Issue (CS:S)
In Counter-Strike: Source the skills which granted the player health upon spawning would set the health prior to the game setting the player's health to 100, which effectively resulted in the player never receiving the benefits from skills that were supposed to give the player additional health upon spawning. With this commit, all issues of the like should have been fixed.

Thanks to Bason for reporting this issue.