Fixed: Raxx - The Oblivion Knight
Changed the description of Raxx - The Oblivion Knight's ultimate skill Enrage's description, as you cannot change the player's size in Counter-Strike: Global Offensive.

Fixed a bug which caused the second skill, Unholy Magic to not work due to the block never being executed.

Added some chat messages when the ultimate is being used.