Fixed: Naix - The Lifestealer
Fixed Naix's Infest skill which wasn't working before, due to player_duck and player_unduck not being available events.
Naix is now using the new wcs_getducking command which will determine whether the player is currently ducking or not.
Also fixed a bug where the infest skill wouldn't trigger if the victim was a bot.