Added: 4 Builder Races
Added the 4 builder races to the current list of races.
These are the four builder races that used to be in a seperate file of their own and disabled by default.
Seemingly I've only had requests on how to add them, and wondering questions as to why they were not included initially, that combined with the introduction of the new commands should make the builder races more compatible with the game mode as you can now actually build properly with them like you used to be able to in CS:S.